batch llm call per file
Optimize by sending a request per file, instead of per line.
As we're now asking the LLM to do more (to change a JSON structure
instead of returning a single line of text), we do a sanity-check
on the result to make sure that it has otherwise followed the
instructions (e.g. hasn't added fields or records in unexpected formats)

Commented out code to attach the file as an attachment.
This is left out because the OpenAI assistants API didn't like .adoc
format.